feat: club detail page for public
