+++ packages/web/src/app/clubs/[id]/page.tsx
@@ -1,60 +1,23 @@
 "use client";
 
-import { useEffect, useState } from "react";
-
-import { RegistrationDeadlineEnum } from "@sparcs-clubs/interface/common/enum/registration.enum";
 import { useParams } from "next/navigation";
 
 import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
-import ClubDetailMainFrame from "@sparcs-clubs/web/features/clubDetails/frames/ClubDetailMainFrame";
+import { useAuth } from "@sparcs-clubs/web/common/providers/AuthContext";
+import ClubDetailPublicFrame from "@sparcs-clubs/web/features/clubDetails/frames/ClubDetailPublicFrame";
+import ClubDetailStudentFrame from "@sparcs-clubs/web/features/clubDetails/frames/ClubDetailStudentFrame";
 import { useGetClubDetail } from "@sparcs-clubs/web/features/clubDetails/services/getClubDetail";
-import { useGetRegistrationTerm } from "@sparcs-clubs/web/features/clubs/services/useGetRegistrationTerm";
 
 const ClubDetail = () => {
   const { id } = useParams();
   const { data, isLoading, isError } = useGetClubDetail(id as string);
-  const {
-    data: termData,
-    isLoading: isLoadingTerm,
-    isError: isErrorTerm,
-  } = useGetRegistrationTerm();
-  const [isRegistrationPeriod, setIsRegistrationPeriod] =
-    useState<boolean>(false);
-
-  useEffect(() => {
-    if (termData) {
-      const now = new Date();
-      const currentEvents = termData.events.filter(
-        event => now >= event.startTerm && now <= event.endTerm,
-      );
-      if (currentEvents.length === 0) {
-        setIsRegistrationPeriod(false);
-        return;
-      }
-      const registrationEvent = currentEvents.filter(
-        event =>
-          event.registrationEventEnumId ===
-          RegistrationDeadlineEnum.StudentRegistrationApplication,
-      );
-      if (registrationEvent.length > 0) {
-        setIsRegistrationPeriod(true);
-      } else {
-        setIsRegistrationPeriod(false);
-      }
-    }
-  }, [termData]);
+  const { isLoggedIn } = useAuth();
 
   return (
-    <AsyncBoundary
-      isLoading={isLoading || isLoadingTerm}
-      isError={isError || isErrorTerm}
-    >
-      {data && (
-        <ClubDetailMainFrame
-          club={data}
-          isRegistrationPeriod={isRegistrationPeriod}
-        />
-      )}
+    <AsyncBoundary isLoading={isLoading} isError={isError}>
+      {isLoggedIn
+        ? data && <ClubDetailStudentFrame club={data} />
+        : data && <ClubDetailPublicFrame club={data} />}{" "}
     </AsyncBoundary>
   );
 };
